Separate timestamp related test from db_basic_test
